+++ src/main/java/cart/config/auth/AuthProviderConfig.java
@@ -1,4 +1,4 @@
-package cart.config;
+package cart.config.auth;
 
 import cart.application.MemberService;
 import org.springframework.context.annotation.Bean;
