+++ src/main/java/cart/config/auth/UserArgumentResolver.java
@@ -1,4 +1,4 @@
-package cart.config;
+package cart.config.auth;
 
 import cart.dto.User;
 import cart.exception.AuthenticationException;
