+++ src/main/java/cart/dao/MemberDao.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.jdbc.core.JdbcTemplate;
