+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -1,6 +1,6 @@
 package cart.dto.response;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 public class ProductResponse {
     private Long id;
