+++ src/main/java/cart/repository/OrderRepository.java
@@ -3,9 +3,9 @@ package cart.repository;
 import cart.dao.MemberDao;
 import cart.dao.OrderDao;
 import cart.dao.OrderItemDao;
-import cart.domain.Member;
-import cart.domain.Order;
-import cart.domain.OrderItem;
+import cart.domain.member.Member;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
 import cart.exception.MemberNotFoundException;
 import cart.exception.OrderNotFoundException;
 import cart.repository.dto.OrderEntity;
