+++ src/test/java/cart/MockAuthProviderConfig.java
@@ -1,7 +1,7 @@
 package cart;
 
-import cart.config.AuthProvider;
-import cart.config.BasicAuthProvider;
+import cart.config.auth.AuthProvider;
+import cart.config.auth.BasicAuthProvider;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
