+++ src/test/java/cart/application/OrderServiceTest.java
@@ -9,10 +9,10 @@ import static org.mockito.BDDMockito.given;
 import cart.dao.CartItemDao;
 import cart.dao.MemberDao;
 import cart.dao.ProductDao;
-import cart.domain.Member;
-import cart.domain.Order;
-import cart.domain.OrderItem;
-import cart.domain.Product;
+import cart.domain.member.Member;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
+import cart.domain.product.Product;
 import cart.dto.request.OrderItemRequest;
 import cart.dto.request.OrderRequest;
 import cart.dto.response.OrderDetailResponse;
