+++ src/test/java/cart/domain/price/PointTest.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.price;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
