+++ src/test/java/cart/domain/product/ProductTest.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.product;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
