+++ src/test/java/cart/domain/product/QuantityTest.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.product;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
