+++ src/test/java/cart/repository/OrderRepositoryTest.java
@@ -6,10 +6,10 @@ import cart.dao.MemberDao;
 import cart.dao.OrderDao;
 import cart.dao.OrderItemDao;
 import cart.dao.ProductDao;
-import cart.domain.Member;
-import cart.domain.Order;
-import cart.domain.OrderItem;
-import cart.domain.Product;
+import cart.domain.member.Member;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
+import cart.domain.product.Product;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
